moved snatch selection to vue
