+++ themes/dark/templates/layouts/main.mako
@@ -116,7 +116,6 @@
         <script type="text/javascript" src="js/home/index.js?${sbPID}"></script>
         <script type="text/javascript" src="js/home/post-process.js?${sbPID}"></script>
         <script type="text/javascript" src="js/home/restart.js?${sbPID}"></script>
-        <script type="text/javascript" src="js/home/snatch-selection.js?${sbPID}"></script>
         <script type="text/javascript" src="js/home/status.js?${sbPID}"></script>
 
         <script type="text/javascript" src="js/manage/failed-downloads.js?${sbPID}"></script>
