+++ themes/light/assets/js/vendors.js
@@ -716,6 +716,18 @@ eval("\n\nvar _extends = Object.assign || function (target) { for (var i = 1; i 
 
 /***/ }),
 
+/***/ "./node_modules/babel-loader/lib/index.js!./node_modules/vue-loader/lib/index.js?!./src/components/snatch-selection.vue?vue&type=script&lang=js&":
+/*!********************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib!./node_modules/vue-loader/lib??vue-loader-options!./src/components/snatch-selection.vue?vue&type=script&lang=js& ***!
+  \********************************************************************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+"use strict";
+eval("/* WEBPACK VAR INJECTION */(function($) {\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };\n\nvar _templateObject = _taggedTemplateLiteral([''], ['']),\n    _templateObject2 = _taggedTemplateLiteral(['Something went wrong in getthing the paramaters from dom. indexerName: ', ',\\n                            seriesId: ', ', season: ', ', episode: ', ''], ['Something went wrong in getthing the paramaters from dom. indexerName: ', ',\\n                            seriesId: ', ', season: ', ', episode: ', '']);\n\nvar _vuex = __webpack_require__(/*! vuex */ \"./node_modules/vuex/dist/vuex.esm.js\");\n\nvar _api = __webpack_require__(/*! ../api */ \"./src/api.js\");\n\nvar _appLink = __webpack_require__(/*! ./app-link.vue */ \"./src/components/app-link.vue\");\n\nvar _appLink2 = _interopRequireDefault(_appLink);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _taggedTemplateLiteral(strings, raw) { return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }\n\nexports.default = {\n    name: 'snatchSelection',\n    template: '#snatch-selection-template',\n    components: {\n        AppLink: _appLink2.default\n    },\n    computed: _extends({}, (0, _vuex.mapState)({\n        shows: function shows(state) {\n            return state.shows.shows;\n        }\n    }), (0, _vuex.mapGetters)(['getShowById']), {\n        indexer: function indexer() {\n            return this.$route.query.indexername;\n        },\n        id: function id() {\n            return this.$route.query.seriesid;\n        },\n        season: function season() {\n            return this.$route.query.season;\n        },\n        episode: function episode() {\n            return this.$route.query.episode;\n        },\n        show: function show() {\n            var indexer = this.indexer,\n                id = this.id,\n                getShowById = this.getShowById,\n                shows = this.shows,\n                $store = this.$store;\n            var defaults = $store.state.defaults;\n\n\n            if (shows.length === 0 || !indexer || !id) {\n                return defaults.show;\n            }\n\n            var show = getShowById({ indexer: indexer, id: id });\n            if (!show) {\n                return defaults.show;\n            }\n\n            return show;\n        }\n    }),\n    mounted: function mounted() {\n        window.addEventListener('load', function () {\n            // Adjust the summary background position and size\n            window.dispatchEvent(new Event('resize'));\n        });\n\n        attachImdbTooltip(); // eslint-disable-line no-undef\n\n        $(document.body).on('click', '.imdbPlot', function (event) {\n            $(event.currentTarget).prev('span').toggle();\n            if ($(event.currentTarget).html() === '..show less') {\n                $(event.currentTarget).html('..show more');\n            } else {\n                $(event.currentTarget).html('..show less');\n            }\n            moveSummaryBackground();\n        });\n\n        // Adjust the summary background position and size on page load and resize\n        function moveSummaryBackground() {\n            var height = $('#summary').height() + 10;\n            var top = $('#summary').offset().top + 5;\n            $('#summaryBackground').height(height);\n            $('#summaryBackground').offset({ top: top, left: 0 });\n            $('#summaryBackground').show();\n        }\n\n        $(window).resize(function () {\n            moveSummaryBackground();\n        });\n\n        var updateSpinner = function updateSpinner(message, showSpinner) {\n            // Get spinner object as needed\n            var spinner = $('#searchNotification');\n            if (showSpinner) {\n                message = '<img id=\"searchingAnim\" src=\"images/loading32' + MEDUSA.config.themeSpinner + '.gif\" height=\"16\" width=\"16\" />&nbsp;' + message;\n            }\n            $(spinner).empty().append(message);\n        };\n\n        // Check the previous status of the history table, for hidden or shown, through the data attribute\n        // data-history-toggle-hidden\n        function toggleHistoryTable() {\n            // eslint-disable-line no-unused-vars\n            // Get previous state which was saved on the wrapper\n            var showOrHide = $('#wrapper').attr('data-history-toggle');\n            $('#historydata').collapse(showOrHide);\n        }\n\n        $.fn.loadContainer = function (path, loadingTxt, errorTxt, callback) {\n            updateSpinner(loadingTxt);\n            $('#manualSearchMeta').load(path + ' #manualSearchMeta meta');\n            $(this).load(path + ' #manualSearchTbody tr', function (response, status) {\n                if (status === 'error') {\n                    updateSpinner(errorTxt, false);\n                }\n                if (typeof callback !== 'undefined') {\n                    callback();\n                }\n            });\n        };\n\n        // Click event for the download button for snatching a result\n        $(document.body).on('click', '.epManualSearch', function (event) {\n            event.preventDefault();\n            var link = event.currentTarget;\n            $(link).children('img').prop('src', 'images/loading16.gif');\n            $.getJSON(event.currentTarget.href, function (data) {\n                if (data.result === 'success') {\n                    $(link).children('img').prop('src', 'images/save.png');\n                } else {\n                    $(link).children('img').prop('src', 'images/no16.png');\n                }\n            });\n        });\n\n        $.fn.generateStars = function () {\n            return this.each(function (index, element) {\n                $(element).html($('<span/>').width($(element).text() * 12));\n            });\n        };\n\n        function initTableSorter(table) {\n            // Nasty hack to re-initialize tablesorter after refresh\n            $(table).tablesorter({\n                widgets: ['saveSort', 'stickyHeaders', 'columnSelector', 'filter'],\n                widgetOptions: {\n                    filter_columnFilters: true, // eslint-disable-line camelcase\n                    filter_hideFilters: true, // eslint-disable-line camelcase\n                    filter_saveFilters: true, // eslint-disable-line camelcase\n                    columnSelector_saveColumns: true, // eslint-disable-line camelcase\n                    columnSelector_layout: '<label><input type=\"checkbox\">{name}</label>', // eslint-disable-line camelcase\n                    columnSelector_mediaquery: false, // eslint-disable-line camelcase\n                    columnSelector_cssChecked: 'checked' // eslint-disable-line camelcase\n                },\n                textExtraction: function () {\n                    return {\n                        // 2: Provider\n                        2: function _(node) {\n                            return $(node).find('img').attr('title');\n                        },\n\n                        // 6: The size column needs an explicit field for the filtering on raw size.\n                        6: function _(node) {\n                            return node.getAttribute('data-size');\n                        },\n\n                        // 9: Published date\n                        9: function _(node) {\n                            return node.getAttribute('data-datetime');\n                        },\n\n                        // 10: Added date\n                        10: function _(node) {\n                            return node.getAttribute('data-datetime');\n                        }\n                    };\n                }(),\n                headers: {\n                    9: { sorter: 'realISODate' }, // Published\n                    10: { sorter: 'realISODate' }, // Added\n                    11: { sorter: false, parser: false // Snatch link\n                    } }\n            });\n        }\n\n        $('.imdbstars').generateStars();\n\n        function checkCacheUpdates(repeat) {\n            var self = this;\n            var pollInterval = 5000;\n            repeat = repeat || true;\n\n            var indexerName = $('meta[data-last-prov-updates]').attr('data-indexer-name');\n            var seriesId = $('meta[data-last-prov-updates]').attr('data-series-id');\n            var season = $('meta[data-last-prov-updates]').attr('data-season');\n            var episode = $('meta[data-last-prov-updates]').attr('data-episode');\n            var data = $('meta[data-last-prov-updates]').data('last-prov-updates');\n            var manualSearchType = $('meta[data-last-prov-updates]').attr('data-manual-search-type');\n\n            var checkParams = [indexerName, seriesId, season, episode].every(function (checkIsTrue) {\n                return checkIsTrue;\n            });\n\n            if (!checkParams) {\n                console.log(''(_templateObject2, indexerName, seriesId, season, episode)(_templateObject));\n                return;\n            }\n\n            var urlParams = '?indexername=' + indexerName + '&seriesid=' + seriesId + '&season=' + season + '&episode=' + episode;\n\n            if (manualSearchType === 'season') {\n                urlParams += '&manual_search_type=' + manualSearchType;\n            }\n\n            if (!$.isNumeric(seriesId) || !$.isNumeric(season) || !$.isNumeric(episode)) {\n                setTimeout(function () {\n                    checkCacheUpdates(true);\n                }, 200);\n            }\n\n            self.refreshResults = function () {\n                // @FIXME: In the current transition phase to Vue, we can't load pages in this way, so use reload instead.\n                /*\n                $('#manualSearchTbody').loadContainer(\n                    'home/snatchSelection' + urlParams,\n                    'Loading new search results...',\n                    'Time out, refresh page to try again',\n                    toggleHistoryTable // This is a callback function\n                );\n                */\n                window.location.reload();\n            };\n\n            $.ajax({\n                url: 'home/manualSearchCheckCache' + urlParams,\n                type: 'GET',\n                data: data,\n                contentType: 'application/json',\n                error: function error() {\n                    // Repeat = false;\n                    console.log('Error occurred!!');\n                    $('.manualSearchButton').removeAttr('disabled');\n                },\n                complete: function complete() {\n                    if (repeat) {\n                        setTimeout(checkCacheUpdates, pollInterval);\n                    }\n                },\n\n                timeout: 15000 // Timeout after 15s\n            }).done(function (data) {\n                // @TODO: Combine the lower if statements\n                if (data === '') {\n                    updateSpinner('Search finished', false);\n                    $('.manualSearchButton').removeAttr('disabled');\n                    repeat = false;\n                }\n\n                if (data.result === 'refresh') {\n                    self.refreshResults();\n                    updateSpinner('Refreshed results...', true);\n                }\n                if (data.result === 'searching') {\n                    // Ep is searched, you will get a results any minute now\n                    pollInterval = 5000;\n                    $('.manualSearchButton').prop('disabled', true);\n                    updateSpinner('The episode is being searched, please wait......', true);\n                }\n                if (data.result === 'queued') {\n                    // Ep is queued, this might take some time to get results\n                    pollInterval = 7000;\n                    $('.manualSearchButton').prop('disabled', true);\n                    updateSpinner('The episode has been queued, because another search is taking place. please wait..', true);\n                }\n                if (data.result === 'finished') {\n                    // Ep search is finished\n                    updateSpinner('Search finished', false);\n                    $('.manualSearchButton').removeAttr('disabled');\n                    repeat = false;\n                    $('#srchresults').trigger('update', true);\n                    $('[datetime]').timeago();\n                }\n                if (data.result === 'error') {\n                    // Ep search is finished but with an error\n                    console.log('Probably tried to call manualSelectCheckCache, while page was being refreshed.');\n                    $('.manualSearchButton').removeAttr('disabled');\n                    repeat = true;\n                }\n            });\n        }\n\n        setTimeout(checkCacheUpdates, 2000);\n\n        // Click event for the reload results and force search buttons\n        $(document.body).on('click', '.manualSearchButton', function (event) {\n            event.preventDefault();\n            $('.manualSearchButton').prop('disabled', true);\n            var indexerName = $('meta[data-last-prov-updates]').attr('data-indexer-name');\n            var seriesId = $('meta[data-last-prov-updates]').attr('data-series-id');\n            var season = $('meta[data-last-prov-updates]').attr('data-season');\n            var episode = $('meta[data-last-prov-updates]').attr('data-episode');\n            var manualSearchType = $('meta[data-last-prov-updates]').attr('data-manual-search-type');\n            var forceSearch = $(event.currentTarget).attr('data-force-search');\n\n            var checkParams = [indexerName, seriesId, season, episode].every(function (checkIsTrue) {\n                return checkIsTrue;\n            });\n\n            if (!checkParams) {\n                console.log(''(_templateObject2, indexerName, seriesId, season, episode)(_templateObject));\n                return;\n            }\n\n            if ($.isNumeric(seriesId) && $.isNumeric(season) && $.isNumeric(episode)) {\n                updateSpinner('Started a forced manual search...', true);\n                $.getJSON('home/snatchSelection', {\n                    indexername: indexerName,\n                    seriesid: seriesId,\n                    season: season,\n                    episode: episode,\n                    manual_search_type: manualSearchType, // eslint-disable-line camelcase\n                    perform_search: forceSearch // eslint-disable-line camelcase\n                });\n                // Force the search, but give the checkCacheUpdates the time to start up a search thread\n                setTimeout(function () {\n                    checkCacheUpdates(true);\n                }, 2000);\n            }\n        });\n\n        // Moved and rewritten this from displayShow. This changes the button when clicked for collapsing/expanding the\n        // \"Show History\" button to show or hide the snatch/download/failed history for a manual searched episode or pack.\n\n        $('#popover').popover({\n            placement: 'bottom',\n            html: true, // Required if content has HTML\n            content: '<div id=\"popover-target\"></div>'\n        }).on('shown.bs.popover', function () {\n            // Bootstrap popover event triggered when the popover opens\n            $.tablesorter.columnSelector.attachTo($('#srchresults'), '#popover-target');\n        });\n\n        $('#btnReset').click(function () {\n            $('#showTable').trigger('saveSortReset') // Clear saved sort\n            .trigger('sortReset'); // Reset current table sort\n            return false;\n        });\n\n        $(function () {\n            initTableSorter('#srchresults');\n            moveSummaryBackground();\n            $('body').on('hide.bs.collapse', '.collapse.toggle', function () {\n                $('#showhistory').text('Show History');\n                $('#wrapper').prop('data-history-toggle', 'hide');\n            });\n            $('body').on('show.bs.collapse', '.collapse.toggle', function () {\n                $('#showhistory').text('Hide History');\n                $('#wrapper').prop('data-history-toggle', 'show');\n            });\n        });\n\n        $(document.body).on('click', '.release-name-ellipses, .release-name-ellipses-toggled', function (event) {\n            var target = $(event.currentTarget);\n\n            if (target.hasClass('release-name-ellipses')) {\n                target.switchClass('release-name-ellipses', 'release-name-ellipses-toggled', 100);\n            } else {\n                target.switchClass('release-name-ellipses-toggled', 'release-name-ellipses', 100);\n            }\n        });\n    }\n};\n/* WEBPACK VAR INJECTION */}.call(this, __webpack_require__(/*! jquery */ \"./node_modules/jquery/dist/jquery.js\")))\n\n//# sourceURL=webpack:///./src/components/snatch-selection.vue?./node_modules/babel-loader/lib!./node_modules/vue-loader/lib??vue-loader-options");
+
+/***/ }),
+
 /***/ "./node_modules/bootstrap/dist/css/bootstrap.min.css":
 /*!***********************************************************!*\
   !*** ./node_modules/bootstrap/dist/css/bootstrap.min.css ***!
@@ -2649,7 +2661,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\nexports.NotFound = exports.ShowSelector = exports.SelectList = exports.ScrollButtons = exports.RootDirs = exports.PlotInfo = exports.NamePattern = exports.Login = exports.LanguageSelect = exports.FileBrowser = exports.DisplayShow = exports.ConfigToggleSlider = exports.ConfigTextboxNumber = exports.ConfigTextbox = exports.Config = exports.Backstretch = exports.Asset = exports.AppLink = exports.AppHeader = exports.AnidbReleaseGroupUi = exports.AddShows = exports.AddRecommended = undefined;\n\nvar _addRecommended = __webpack_require__(/*! ./add-recommended.vue */ \"./src/components/add-recommended.vue\");\n\nvar _addRecommended2 = _interopRequireDefault(_addRecommended);\n\nvar _addShows = __webpack_require__(/*! ./add-shows.vue */ \"./src/components/add-shows.vue\");\n\nvar _addShows2 = _interopRequireDefault(_addShows);\n\nvar _anidbReleaseGroupUi = __webpack_require__(/*! ./anidb-release-group-ui.vue */ \"./src/components/anidb-release-group-ui.vue\");\n\nvar _anidbReleaseGroupUi2 = _interopRequireDefault(_anidbReleaseGroupUi);\n\nvar _appHeader = __webpack_require__(/*! ./app-header.vue */ \"./src/components/app-header.vue\");\n\nvar _appHeader2 = _interopRequireDefault(_appHeader);\n\nvar _appLink = __webpack_require__(/*! ./app-link.vue */ \"./src/components/app-link.vue\");\n\nvar _appLink2 = _interopRequireDefault(_appLink);\n\nvar _asset = __webpack_require__(/*! ./asset.vue */ \"./src/components/asset.vue\");\n\nvar _asset2 = _interopRequireDefault(_asset);\n\nvar _backstretch = __webpack_require__(/*! ./backstretch.vue */ \"./src/components/backstretch.vue\");\n\nvar _backstretch2 = _interopRequireDefault(_backstretch);\n\nvar _config = __webpack_require__(/*! ./config.vue */ \"./src/components/config.vue\");\n\nvar _config2 = _interopRequireDefault(_config);\n\nvar _configTextbox = __webpack_require__(/*! ./config-textbox.vue */ \"./src/components/config-textbox.vue\");\n\nvar _configTextbox2 = _interopRequireDefault(_configTextbox);\n\nvar _configTextboxNumber = __webpack_require__(/*! ./config-textbox-number.vue */ \"./src/components/config-textbox-number.vue\");\n\nvar _configTextboxNumber2 = _interopRequireDefault(_configTextboxNumber);\n\nvar _configToggleSlider = __webpack_require__(/*! ./config-toggle-slider.vue */ \"./src/components/config-toggle-slider.vue\");\n\nvar _configToggleSlider2 = _interopRequireDefault(_configToggleSlider);\n\nvar _displayShow = __webpack_require__(/*! ./display-show.vue */ \"./src/components/display-show.vue\");\n\nvar _displayShow2 = _interopRequireDefault(_displayShow);\n\nvar _fileBrowser = __webpack_require__(/*! ./file-browser.vue */ \"./src/components/file-browser.vue\");\n\nvar _fileBrowser2 = _interopRequireDefault(_fileBrowser);\n\nvar _languageSelect = __webpack_require__(/*! ./language-select.vue */ \"./src/components/language-select.vue\");\n\nvar _languageSelect2 = _interopRequireDefault(_languageSelect);\n\nvar _login = __webpack_require__(/*! ./login.vue */ \"./src/components/login.vue\");\n\nvar _login2 = _interopRequireDefault(_login);\n\nvar _namePattern = __webpack_require__(/*! ./name-pattern.vue */ \"./src/components/name-pattern.vue\");\n\nvar _namePattern2 = _interopRequireDefault(_namePattern);\n\nvar _plotInfo = __webpack_require__(/*! ./plot-info.vue */ \"./src/components/plot-info.vue\");\n\nvar _plotInfo2 = _interopRequireDefault(_plotInfo);\n\nvar _rootDirs = __webpack_require__(/*! ./root-dirs.vue */ \"./src/components/root-dirs.vue\");\n\nvar _rootDirs2 = _interopRequireDefault(_rootDirs);\n\nvar _scrollButtons = __webpack_require__(/*! ./scroll-buttons.vue */ \"./src/components/scroll-buttons.vue\");\n\nvar _scrollButtons2 = _interopRequireDefault(_scrollButtons);\n\nvar _selectList = __webpack_require__(/*! ./select-list.vue */ \"./src/components/select-list.vue\");\n\nvar _selectList2 = _interopRequireDefault(_selectList);\n\nvar _showSelector = __webpack_require__(/*! ./show-selector.vue */ \"./src/components/show-selector.vue\");\n\nvar _showSelector2 = _interopRequireDefault(_showSelector);\n\nvar _http = __webpack_require__(/*! ./http */ \"./src/components/http/index.js\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.AddRecommended = _addRecommended2.default;\nexports.AddShows = _addShows2.default;\nexports.AnidbReleaseGroupUi = _anidbReleaseGroupUi2.default;\nexports.AppHeader = _appHeader2.default;\nexports.AppLink = _appLink2.default;\nexports.Asset = _asset2.default;\nexports.Backstretch = _backstretch2.default;\nexports.Config = _config2.default;\nexports.ConfigTextbox = _configTextbox2.default;\nexports.ConfigTextboxNumber = _configTextboxNumber2.default;\nexports.ConfigToggleSlider = _configToggleSlider2.default;\nexports.DisplayShow = _displayShow2.default;\nexports.FileBrowser = _fileBrowser2.default;\nexports.LanguageSelect = _languageSelect2.default;\nexports.Login = _login2.default;\nexports.NamePattern = _namePattern2.default;\nexports.PlotInfo = _plotInfo2.default;\nexports.RootDirs = _rootDirs2.default;\nexports.ScrollButtons = _scrollButtons2.default;\nexports.SelectList = _selectList2.default;\nexports.ShowSelector = _showSelector2.default;\nexports.NotFound = _http.NotFound;\n\n//# sourceURL=webpack:///./src/components/index.js?");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\nexports.NotFound = exports.SnatchSelection = exports.ShowSelector = exports.SelectList = exports.ScrollButtons = exports.RootDirs = exports.PlotInfo = exports.NamePattern = exports.Login = exports.LanguageSelect = exports.FileBrowser = exports.DisplayShow = exports.ConfigToggleSlider = exports.ConfigTextboxNumber = exports.ConfigTextbox = exports.Config = exports.Backstretch = exports.Asset = exports.AppLink = exports.AppHeader = exports.AnidbReleaseGroupUi = exports.AddShows = exports.AddRecommended = undefined;\n\nvar _addRecommended = __webpack_require__(/*! ./add-recommended.vue */ \"./src/components/add-recommended.vue\");\n\nvar _addRecommended2 = _interopRequireDefault(_addRecommended);\n\nvar _addShows = __webpack_require__(/*! ./add-shows.vue */ \"./src/components/add-shows.vue\");\n\nvar _addShows2 = _interopRequireDefault(_addShows);\n\nvar _anidbReleaseGroupUi = __webpack_require__(/*! ./anidb-release-group-ui.vue */ \"./src/components/anidb-release-group-ui.vue\");\n\nvar _anidbReleaseGroupUi2 = _interopRequireDefault(_anidbReleaseGroupUi);\n\nvar _appHeader = __webpack_require__(/*! ./app-header.vue */ \"./src/components/app-header.vue\");\n\nvar _appHeader2 = _interopRequireDefault(_appHeader);\n\nvar _appLink = __webpack_require__(/*! ./app-link.vue */ \"./src/components/app-link.vue\");\n\nvar _appLink2 = _interopRequireDefault(_appLink);\n\nvar _asset = __webpack_require__(/*! ./asset.vue */ \"./src/components/asset.vue\");\n\nvar _asset2 = _interopRequireDefault(_asset);\n\nvar _backstretch = __webpack_require__(/*! ./backstretch.vue */ \"./src/components/backstretch.vue\");\n\nvar _backstretch2 = _interopRequireDefault(_backstretch);\n\nvar _config = __webpack_require__(/*! ./config.vue */ \"./src/components/config.vue\");\n\nvar _config2 = _interopRequireDefault(_config);\n\nvar _configTextbox = __webpack_require__(/*! ./config-textbox.vue */ \"./src/components/config-textbox.vue\");\n\nvar _configTextbox2 = _interopRequireDefault(_configTextbox);\n\nvar _configTextboxNumber = __webpack_require__(/*! ./config-textbox-number.vue */ \"./src/components/config-textbox-number.vue\");\n\nvar _configTextboxNumber2 = _interopRequireDefault(_configTextboxNumber);\n\nvar _configToggleSlider = __webpack_require__(/*! ./config-toggle-slider.vue */ \"./src/components/config-toggle-slider.vue\");\n\nvar _configToggleSlider2 = _interopRequireDefault(_configToggleSlider);\n\nvar _displayShow = __webpack_require__(/*! ./display-show.vue */ \"./src/components/display-show.vue\");\n\nvar _displayShow2 = _interopRequireDefault(_displayShow);\n\nvar _fileBrowser = __webpack_require__(/*! ./file-browser.vue */ \"./src/components/file-browser.vue\");\n\nvar _fileBrowser2 = _interopRequireDefault(_fileBrowser);\n\nvar _languageSelect = __webpack_require__(/*! ./language-select.vue */ \"./src/components/language-select.vue\");\n\nvar _languageSelect2 = _interopRequireDefault(_languageSelect);\n\nvar _login = __webpack_require__(/*! ./login.vue */ \"./src/components/login.vue\");\n\nvar _login2 = _interopRequireDefault(_login);\n\nvar _namePattern = __webpack_require__(/*! ./name-pattern.vue */ \"./src/components/name-pattern.vue\");\n\nvar _namePattern2 = _interopRequireDefault(_namePattern);\n\nvar _plotInfo = __webpack_require__(/*! ./plot-info.vue */ \"./src/components/plot-info.vue\");\n\nvar _plotInfo2 = _interopRequireDefault(_plotInfo);\n\nvar _rootDirs = __webpack_require__(/*! ./root-dirs.vue */ \"./src/components/root-dirs.vue\");\n\nvar _rootDirs2 = _interopRequireDefault(_rootDirs);\n\nvar _scrollButtons = __webpack_require__(/*! ./scroll-buttons.vue */ \"./src/components/scroll-buttons.vue\");\n\nvar _scrollButtons2 = _interopRequireDefault(_scrollButtons);\n\nvar _selectList = __webpack_require__(/*! ./select-list.vue */ \"./src/components/select-list.vue\");\n\nvar _selectList2 = _interopRequireDefault(_selectList);\n\nvar _showSelector = __webpack_require__(/*! ./show-selector.vue */ \"./src/components/show-selector.vue\");\n\nvar _showSelector2 = _interopRequireDefault(_showSelector);\n\nvar _snatchSelection = __webpack_require__(/*! ./snatch-selection.vue */ \"./src/components/snatch-selection.vue\");\n\nvar _snatchSelection2 = _interopRequireDefault(_snatchSelection);\n\nvar _http = __webpack_require__(/*! ./http */ \"./src/components/http/index.js\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.AddRecommended = _addRecommended2.default;\nexports.AddShows = _addShows2.default;\nexports.AnidbReleaseGroupUi = _anidbReleaseGroupUi2.default;\nexports.AppHeader = _appHeader2.default;\nexports.AppLink = _appLink2.default;\nexports.Asset = _asset2.default;\nexports.Backstretch = _backstretch2.default;\nexports.Config = _config2.default;\nexports.ConfigTextbox = _configTextbox2.default;\nexports.ConfigTextboxNumber = _configTextboxNumber2.default;\nexports.ConfigToggleSlider = _configToggleSlider2.default;\nexports.DisplayShow = _displayShow2.default;\nexports.FileBrowser = _fileBrowser2.default;\nexports.LanguageSelect = _languageSelect2.default;\nexports.Login = _login2.default;\nexports.NamePattern = _namePattern2.default;\nexports.PlotInfo = _plotInfo2.default;\nexports.RootDirs = _rootDirs2.default;\nexports.ScrollButtons = _scrollButtons2.default;\nexports.SelectList = _selectList2.default;\nexports.ShowSelector = _showSelector2.default;\nexports.SnatchSelection = _snatchSelection2.default;\nexports.NotFound = _http.NotFound;\n\n//# sourceURL=webpack:///./src/components/index.js?");
 
 /***/ }),
 
@@ -3034,6 +3046,30 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _nod
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_show_selector_vue_vue_type_template_id_1f9d7d4f___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../node_modules/vue-loader/lib??vue-loader-options!./show-selector.vue?vue&type=template&id=1f9d7d4f& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./src/components/show-selector.vue?vue&type=template&id=1f9d7d4f&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_show_selector_vue_vue_type_template_id_1f9d7d4f___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_show_selector_vue_vue_type_template_id_1f9d7d4f___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./src/components/show-selector.vue?");
+
+/***/ }),
+
+/***/ "./src/components/snatch-selection.vue":
+/*!*********************************************!*\
+  !*** ./src/components/snatch-selection.vue ***!
+  \*********************************************/
+/*! no static exports found */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./snatch-selection.vue?vue&type=script&lang=js& */ \"./src/components/snatch-selection.vue?vue&type=script&lang=js&\");\n/* harmony reexport (unknown) */ for(var __WEBPACK_IMPORT_KEY__ in _snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__) if(__WEBPACK_IMPORT_KEY__ !== 'default') (function(key) { __webpack_require__.d(__webpack_exports__, key, function() { return _snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[key]; }) }(__WEBPACK_IMPORT_KEY__));\n/* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ \"./node_modules/vue-loader/lib/runtime/componentNormalizer.js\");\nvar render, staticRenderFns\n\n\n\n\n/* normalize component */\n\nvar component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"])(\n  _snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[\"default\"],\n  render,\n  staticRenderFns,\n  false,\n  null,\n  null,\n  null\n  \n)\n\n/* hot reload */\nif (false) { var api; }\ncomponent.options.__file = \"src/components/snatch-selection.vue\"\n/* harmony default export */ __webpack_exports__[\"default\"] = (component.exports);\n\n//# sourceURL=webpack:///./src/components/snatch-selection.vue?");
+
+/***/ }),
+
+/***/ "./src/components/snatch-selection.vue?vue&type=script&lang=js&":
+/*!**********************************************************************!*\
+  !*** ./src/components/snatch-selection.vue?vue&type=script&lang=js& ***!
+  \**********************************************************************/
+/*! no static exports found */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_node_modules_vue_loader_lib_index_js_vue_loader_options_snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../node_modules/babel-loader/lib!../../node_modules/vue-loader/lib??vue-loader-options!./snatch-selection.vue?vue&type=script&lang=js& */ \"./node_modules/babel-loader/lib/index.js!./node_modules/vue-loader/lib/index.js?!./src/components/snatch-selection.vue?vue&type=script&lang=js&\");\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_node_modules_vue_loader_lib_index_js_vue_loader_options_snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_babel_loader_lib_index_js_node_modules_vue_loader_lib_index_js_vue_loader_options_snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__);\n/* harmony reexport (unknown) */ for(var __WEBPACK_IMPORT_KEY__ in _node_modules_babel_loader_lib_index_js_node_modules_vue_loader_lib_index_js_vue_loader_options_snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__) if(__WEBPACK_IMPORT_KEY__ !== 'default') (function(key) { __webpack_require__.d(__webpack_exports__, key, function() { return _node_modules_babel_loader_lib_index_js_node_modules_vue_loader_lib_index_js_vue_loader_options_snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[key]; }) }(__WEBPACK_IMPORT_KEY__));\n /* harmony default export */ __webpack_exports__[\"default\"] = (_node_modules_babel_loader_lib_index_js_node_modules_vue_loader_lib_index_js_vue_loader_options_snatch_selection_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0___default.a); \n\n//# sourceURL=webpack:///./src/components/snatch-selection.vue?");
 
 /***/ }),
